check for server first, fix chmod chgrp so others can pull on tb-driver

+++ mh_pull_hack.sh
@@ -3,6 +3,10 @@
 cd /opt/enterprise/dists
 wget -O LATEST.new http://pluto.puppetlabs.lan/dists/LATEST 2>/dev/null 
 
+if [ $? -ne 0 ] ; then
+  echo "we lost pluto"
+  exit
+fi
 
 if [ ` cat LATEST ` == ` cat LATEST.new ` ] ; then
   rm LATEST.new
@@ -17,5 +21,5 @@ wget -O - http://pluto.puppetlabs.lan/dists 2>/dev/null | perl -e 'while (<>){ p
 do wget -nc http://pluto.puppetlabs.lan/dists/${tarball} ;
 done
 
-# chgrp developers *
-chmod 0664 LATEST *.tar
+chgrp developers LATEST puppet-enterprise-*.tar
+chmod 0664 LATEST puppet-enterprise-*.tar
